% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_mod_exports.R
\name{default_export_plugin}
\alias{default_export_plugin}
\alias{mod_export_ui}
\alias{mod_export_server}
\title{xxxx}
\usage{
mod_export_ui(id)

mod_export_server(id, object)
}
\arguments{
\item{id}{xxxx}

\item{object}{xxx}
}
\description{
xx
}
