% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineA.R
\name{PipelineA_conf}
\alias{PipelineA_conf}
\title{Shiny example module `Pipeline A`}
\usage{
PipelineA_conf()
}
\description{
This module contains the configuration information for the corresponding pipeline.
It is called by the nav_pipeline module of the package MagellanNTK
This documentation is for developpers who want to create their own pipelines nor processes
to be managed with `MagellanNTK`.
}
