% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineA_Process1.R
\name{example_module_process1}
\alias{example_module_process1}
\alias{PipelineA_Process1_conf}
\alias{PipelineA_Process1_ui}
\alias{PipelineA_Process1_server}
\title{Shiny example process module.}
\usage{
PipelineA_Process1_conf()

PipelineA_Process1_ui(id)

PipelineA_Process1_server(
  id,
  dataIn = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     FALSE
 }),
  steps.status = reactive({
     NULL
 }),
  current.pos = reactive({
     1
 }),
  path = NULL
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{The dataset}

\item{steps.enabled}{A vector of boolean which has the same length of the steps
of the pipeline. This information is used to enable/disable the widgets. It is not
a communication variable between the caller and this module, thus there is no
corresponding output variable}

\item{remoteReset}{It is a remote command to reset the module. A boolean that
indicates is the pipeline has been reseted by a program of higher level
Basically, it is the program which has called this module}

\item{steps.status}{xxx}

\item{current.pos}{xxx}

\item{path}{xxx}
}
\description{
This module contains the configuration informations for the corresponding pipeline.
It is called by the nav_pipeline module of the package MagellanNTK

The name of the server and ui functions are formatted with keywords separated by '_', as follows:
* first string `mod`: indicates that it is a Shiny module
* `pipeline name` is the name of the pipeline to which the process belongs
* `process name` is the name of the process itself

This convention is important because MagellanNTK call the different
server and ui functions by building dynamically their name.

In this example, `PipelineA_ProcessA_ui()` and `PipelineA_ProcessA_server()` define
the code for the process `ProcessA` which is part of the pipeline called `PipelineA`.
}
\author{
Samuel Wieczorek
}
