% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineA_Process2.R
\name{PipelineA_Process2_conf}
\alias{PipelineA_Process2_conf}
\title{Shiny example process module.}
\usage{
PipelineA_Process2_conf()
}
\description{
This module contains the configuration informations for the corresponding pipeline.
It is called by the nav_pipeline module of the package MagellanNTK

The name of the server and ui functions are formatted with keywords separated by '_', as follows:
* first string `mod`: indicates that it is a Shiny module
* `pipeline name` is the name of the pipeline to which the process belongs
* `process name` is the name of the process itself

This convention is important because MagellanNTK call the different
server and ui functions by building dynamically their name.

In this example, `PipelineA_ProcessA_ui()` and `PipelineA_ProcessA_server()` define
the code for the process `ProcessA` which is part of the pipeline called `PipelineA`.
}
