% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mod-foo}
\alias{mod-foo}
\title{module foo}
\description{
xxxxx
}
\examples{
if (interactive()){
data(ft_na)
ui <- foo_ui('query')

server <- function(input, output, session) {
  
  rv <- reactiveValues(
    res = NULL
  )
  ll.tags <- c('None' = 'None', 
  qMetadata.def(typeDataset(ft_na[[1]]))$node)
  
  rv$res <- foo_server('query', 
  reset = reactive({NULL}),
  is.enabled = reactive({NULL})
  )
  
  observeEvent(rv$res$dataOut()$value, ignoreNULL = TRUE, ignoreInit = TRUE, {
    print(rv$res$dataOut()$value)
    print(rv$res$dataOut()$widgets)
  })
}

shinyApp(ui=ui, server=server)
}
}
