% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_mod_openfile.R
\name{readFile}
\alias{readFile}
\title{Read a file}
\usage{
readFile(name, path)
}
\arguments{
\item{name}{xxx}

\item{path}{xxx}
}
\description{
This function is the implementation of the empty readFiles
function (used in MagellantNTK)
}
